Update README
We need to move each component into its own repository.

Why? We want to be able to make incremental changes to the collector
plugins (gems), and we want the Rails app to be able to use those
incremental improvements. You typically accomplish this by setting the
Gemfile in your Rails app to point to the Git repo for the gem.
For example:

  gem "stratify-twitter", :git =>
    "git://github.com/jasonrudolph/stratify-twitter.git"

This approach is not possible as long as all of the gems are stored
in a single repository.

As long as all of the gems are stored in a single repository, you have
to publish a new version of the gem if you want it to be accessible to
the Rails app. Since we don't want to release a new version of the gem
for every small improvement, we soon need to separate the gems out into
their own repositories.

We should consider using Repo to manage the collection of repositories:

http://en.wikipedia.org/wiki/Repo_(script)